update based on Damien comments
