+++ sync/infrahub-sync/infrahub_sync/adapters/netbox.py
@@ -17,10 +17,6 @@ from infrahub_sync import (
 
 from .utils import get_value
 
-# if TYPE_CHECKING:
-#     from pynetbox.core.response import Record as NetboxRecord
-#     from pynetbox.core.response import RecordSet as NetboxRecordSet
-
 
 class NetboxAdapter(DiffSyncMixin, Adapter):
     type = "Netbox"
@@ -154,31 +150,3 @@ class NetboxModel(DiffSyncModelMixin, DiffSyncModel):
     def update(self, attrs):
         # TODO
         return super().update(attrs)
-
-    @classmethod
-    def filter_records(cls, records: list[dict], schema_mapping: SchemaMappingModel) -> list[dict]:
-        """
-        Apply filters to the records based on the schema mapping configuration.
-        """
-        filters = schema_mapping.filters or []
-        if not filters:
-            return records
-        filtered_records = []
-        for record in records:
-            if cls.apply_filters(item=record, filters=filters):
-                filtered_records.append(record)
-        return filtered_records
-
-    @classmethod
-    def transform_records(cls, records: list[dict], schema_mapping: SchemaMappingModel) -> list[dict]:
-        """
-        Apply transformations to the records based on the schema mapping configuration.
-        """
-        transforms = schema_mapping.transforms or []
-        if not transforms:
-            return records
-        transformed_records = []
-        for record in records:
-            transformed_record = cls.apply_transforms(item=record, transforms=transforms)
-            transformed_records.append(transformed_record)
-        return transformed_records
